pyproject.toml
[project]
name = "elt-project"
version = "0.1.0"
description = "BigDataTools Implementation"
authors = [
    {name = "David Chehet"}
]
readme = "README.md"
requires-python = "3.11.12"
dependencies = [
    "cassandra-driver (>=3.29.2,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "uuid (>=1.30,<2.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "geopandas (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
